backend/requirements.txt
asgiref==3.4.1
black==21.11b1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.0
django-cors-headers==3.10.0
django-environ==0.8.1
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.7
pytz==2021.3
regex==2021.11.10
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tomli==1.2.2
typing_extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==6.0.0
